portScannerService/requirements.txt
python-nmap==0.7.1
types-python-nmap==0.7.1.20240106
